Removed 2 commented out methods after reviewing with Joe. Fixed error/ACK bug where an ACK wasnt being sent. Removed readRequest(fileName) method.
the run() method now acts as a controller that calls methods in its class
in the order it desires instead of calling a single method that calls other
methods in a chain. This way, all method calls are in a clear order in
one place.

Created DEFAULT_PATH constant containing "SERVERDATA/" as the root folder
for server data.