Numerous Changes in Client, Error Sim and Server
Current tests completed:

Two transfer cycles were tested (hence DATA 1/2 ACK 1/2).

Read:
  Lose Packet:
    RRQ, DATA 1, DATA 2, ACK 1 all work
    ACK 2 does not work*
Write:
  Lose Packet:
    WRQ, DATA 1, DATA 2 all work
    ACK 1, ACK 2 do not work*

*In all tests, the file was read and written successfully, but sometimes with bugs that cause phantom requests where the server and error sim send things back and forth. This causes the need to manually terminate the error sim since there is no timeout.

Client:
Client now times out, retransmits packets and then waits on a response in that order. Packets are also retransmitted if a packet received is of unexpected type or data and then waits on a response. Client quits trying to receive if it times out three times.

Error Sim thread:
Fixed bug where the first data packet would be passed on and not destroyed. (Added another check earlier on)
Fixed bug where simulator would destroy the last data packet, receive one more packet and then quit since it assumes that the last data packet has been sent even though it destroyed it. Simulator now checks to see if it should be waiting for more packets if the last packet it destroyed was the last data packet.

Server:
Server now times out, retransmits packets and then waits on a response in that order. Packets are also retransmitted if a packet received is of unexpected type or data and then waits on a response. Server quits trying to receive if it times out three times.

Bugs: Threads are restarted because of phantom requests being sent out.