Fixed write request loop in client
Bug: client would try to get an acknowledgement when it shouldn't expect one.
Fix: Deleted writeRequest(filename) method and placed it into sendData(fileText)
to be able to send a packet first and then receive an ack (switching the order
since it is the opposite of a RRQ).

In this commit, most of all iteration 3 appears to be working. Finding big errors
is moderately difficult. Submitting this iteration.