updated file transfer to support more than just text files
I removed some unnecessary string conversions that resulted in anything that is not a .txt file being corrupted. You now should be able to transfer any file type now, provided that the file is not too large for the system to handle.

ex. trying to transfer a .png doesn't work, because the files are too large, (MBs) but transferring a  .jpg works fine because they're much smaller (10s of KBs)

Confirmed successful transfers of formats txt, jpg, pdf, docx, java