Error Code 4 Fix (See details!)
The DATA portion of a packet can now be altered. The problem now is HOW
to alter the data to make it invalid. The packet size can't exceed 516
because the socket ends up trimming the packet so that it'll fit through.
I have yet to find an illegal character to feed it, since writing something
like -9999 byte just wraps around to -15.

Suggestions needed.

Also, Client does not send error packet code 4 to the server. Not sure
if this is supposed to be this way.