fixed bug with saving data from invalid TID
In the previous version, when a DATA packet came in from an unexpected TID, that data was still saved to the file buffer despite being recognized as invalid. This has been fixed for both the client and server. Now, data that is received is only saved if it comes from a valid TID